Updated both public keys
I changed the expire times so they never expire. Most people would think it a hassle to hunt down a new copy that has the newer dates anyhow and not everyone has PGP set up to automatically update since not everyone is on keyservers. This non expiring key makes sense.